requirements.txt
Pillow==9.4.0
Pillow==11.0.0
pyTelegramBotAPI==4.10.0
pyTelegramBotAPI==4.23.0
pytesseract==0.3.10
python-dotenv==1.0.1
Requests==2.32.3
vkbottle==4.4.2
